backend/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
resend==0.8.0
requests==2.31.0
flask-cors==4.0.0
langchain==0.1.0
langchain-openai==0.0.5
openai==1.6.1
pydantic==2.5.0
gunicorn==21.2.0
